+++ engine/src/net/codec/bson_codec.hpp
@@ -16,7 +16,7 @@ public:
     /* 解码数据包
      * return: <0 error,otherwise the number of parameter push to stack
      */
-    int32_t decode(lua_State *L, const char *buffer, int32_t len,
+    int32_t decode(lua_State *L, const char *buffer, size_t len,
                    const CmdCfg *cfg);
     /* 编码数据包
      * return: <0 error,otherwise the length of buffer
