+++ engine/src/net/codec/codec.hpp
@@ -28,7 +28,7 @@ public:
     /* 解码数据包
      * return: <0 error,otherwise the number of parameter push to stack
      */
-    virtual int32_t decode(lua_State *L, const char *buffer, int32_t len,
+    virtual int32_t decode(lua_State *L, const char *buffer, size_t len,
                            const CmdCfg *cfg) = 0;
     /* 编码数据包
      * return: <0 error,otherwise the length of buffer
